+++ pluginmanager/flusher_wrapper_v2.go
@@ -1,0 +1,33 @@
+// Copyright 2021 iLogtail Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package pluginmanager
+
+import (
+	"github.com/alibaba/ilogtail/pkg/models"
+	"github.com/alibaba/ilogtail/pkg/pipeline"
+)
+
+type FlusherWrapperV2 struct {
+	CommonWrapper
+	Flusher pipeline.FlusherV2
+}
+
+func (wrapper *FlusherWrapperV2) Init(name string) error {
+	return wrapper.Flusher.Init(wrapper.Config.Context)
+}
+
+func (wrapper *FlusherWrapperV2) Export(pipelineGroupEvents []*models.PipelineGroupEvents, pipelineContext pipeline.PipelineContext) error {
+	return wrapper.Flusher.Export(pipelineGroupEvents, pipelineContext)
+}
