+++ pluginmanager/processor_wrapper_v2.go
@@ -1,0 +1,34 @@
+// Copyright 2021 iLogtail Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package pluginmanager
+
+import (
+	"github.com/alibaba/ilogtail/pkg/models"
+	"github.com/alibaba/ilogtail/pkg/pipeline"
+)
+
+
+type ProcessorWrapperV2 struct {
+	ProcessorWrapper
+	Processor pipeline.ProcessorV2
+}
+
+func (wrapper *ProcessorWrapperV2) Init(name string) error {
+	return wrapper.Processor.Init(wrapper.Config.Context)
+}
+
+func (wrapper *ProcessorWrapperV2) Process(in *models.PipelineGroupEvents, context pipeline.PipelineContext) {
+	wrapper.Processor.Process(in, context)
+}
